# docs/openapi.yaml
openapi: 3.1.0
info:
  title: KnowledgeOS API
  version: 1.0.0
  summary: Полный REST API KnowledgeOS
  description: |
    Единая OpenAPI-спецификация для backend API KnowledgeOS.

    Файл описывает все маршруты, объявленные в `backend/internal/handler/router.go`:
    авторизацию, базу знаний, комментарии, ссылки, поиск, пользователей,
    компании, настройки бота, RAG, чат, синхронизацию, импорт/экспорт и backup.

    Все JSON-ответы backend, кроме backup snapshot, возвращаются в стандартной
    обёртке:

    - успешный одиночный ответ: `{ "data": ... }`;
    - список: `{ "data": [...], "total": 123 }`;
    - ошибка: `{ "error": "message" }`.

    Protected JWT-методы работают в tenant-контексте. Для обычного пользователя
    компания берётся из JWT. Пользователь с ролью `superadmin` обязан передать
    заголовок `X-Company-ID` для tenant-scoped методов; без выбранной компании
    backend возвращает `403`. Глобальные методы `/admin/companies*` доступны
    superadmin без tenant-контекста.
servers:
  - url: http://localhost:8080/api/v1
    description: Локальный backend
  - url: /api/v1
    description: Относительный путь для frontend/proxy
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Обновить пару токенов
      description: |
        Принимает refresh token, отзывает его и выдаёт новую пару токенов.
        Метод нужен frontend-клиенту для бесшовного продления сессии.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Новые токены.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /webhooks/{channel}/{company_id}:
    post:
      tags: [Channels]
      operationId: handleChannelWebhook
      summary: Принять webhook канала
      description: |
        Public endpoint без JWT для входящих обновлений Telegram, MAX и VK.
        Tenant выбирается по `company_id` в URL, а подлинность проверяется
        channel-specific secret из `tenant_secrets.{channel}.metadata`.
      security: []
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
            enum: [telegram, max, vk]
        - name: company_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Обновление принято или проигнорировано без retry.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Бот или канал выключен.
...
  /admin/bot/channels/{channel}/webhook:
    post:
      tags: [Channels]
      operationId: registerChannelWebhook
      summary: Зарегистрировать webhook канала
      description: |
        Регистрирует (или перерегистрирует) webhook у провайдера канала, используя
        сохранённый token и metadata. Поддерживается для Telegram и MAX. Возвращает
        `registered=false`, если не заполнены обязательные поля для регистрации.
      security:
        - bearerAuth: []
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
            enum: [telegram, max, vk]
      responses:
        '200':
          description: Результат регистрации webhook.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      registered:
                        type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/companies:
    get:
      tags: [Companies]
      operationId: listCompanies
      summary: Получить компании
      description: Список компаний платформы. Доступно только `superadmin`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit50'
      responses:
        '200':
          description: Список компаний.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyListEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Companies]
      operationId: createCompany
      summary: Создать компанию
      description: Создаёт компанию и sync sequence. Доступно только `superadmin`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyInput'
      responses:
        '201':
          description: Созданная компания.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /admin/companies/{id}:
    get:
      tags: [Companies]
      operationId: getCompany
      summary: Получить компанию
      description: Возвращает компанию по UUID. Доступно только `superadmin`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Компания.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Companies]
      operationId: updateCompany
      summary: Обновить компанию
      description: Обновляет компанию. Доступно только `superadmin`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyInput'
      responses:
        '200':
          description: Обновлённая компания.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      tags: [Companies]
      operationId: deleteCompany
      summary: Удалить компанию
      description: Удаляет компанию. Доступно только `superadmin`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '204':
          description: Удалено.
        '400':
          $ref: '#/components/responses/BadRequest'
  /admin/companies/{id}/admin:
    post:
      tags: [Companies, Users]
      operationId: createCompanyAdmin
      summary: Создать администратора компании
      description: |
        Создаёт пользователя с ролью `admin` в выбранной компании. Используйте
        после создания компании, чтобы выдать ей первого администратора.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyAdminRequest'
      responses:
        '201':
          description: Созданный admin user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token из `/auth/login` или `/auth/refresh`.
    syncApiKey:
      type: http
      scheme: bearer
      description: API key из таблицы `api_keys`, используется только `/sync/push` и `/sync/pull`.
    backupApiKey:
      type: http
      scheme: bearer
      description: Значение `BACKUP_API_KEY`, используется только `/backup/snapshot`.
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: UUID сущности.
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: Номер страницы, начиная с 1.
    Limit50:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, default: 50 }
      description: Размер страницы.
...
  responses:
...
    BadRequest:
      description: Некорректный запрос.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error: invalid request body
...
    Unauthorized:
      description: Не передан или невалиден token/API key.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error: invalid token
    Forbidden:
      description: Недостаточно прав.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error: insufficient permissions
    NotFound:
      description: Сущность не найдена или недоступна в текущем tenant.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error: not found
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LoginResponseEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/LoginResponse'
    LoginResponse:
      type: object
      required: [access_token, refresh_token, expires_at, user]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_at:
          type: integer
          format: int64
          description: Unix timestamp expiration для access token.
        user:
          $ref: '#/components/schemas/LoginUser'
    LoginUser:
      type: object
      required: [id, email, role]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        company_id: { type: string, format: uuid }
    Role:
      type: string
      enum: [superadmin, admin, editor, viewer]
...
    BaseModel:
      type: object
      required: [id, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Company:
      allOf:
        - $ref: '#/components/schemas/BaseModel'
        - type: object
          required: [name, tier]
          properties:
            name: { type: string }
            tier: { type: string, default: local }
    CompanyInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        tier: { type: string, default: local }
    User:
      allOf:
        - $ref: '#/components/schemas/BaseModel'
        - type: object
          required: [email, role, is_active]
          properties:
            company_id: { type: string, format: uuid }
            email: { type: string, format: email }
            role: { $ref: '#/components/schemas/Role' }
            is_active: { type: boolean }
...
    CreateCompanyAdminRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
...
    CompanyEnvelope:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Company' }
    CompanyListEnvelope:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Company' }
        total: { type: integer, format: int64 }
    UserEnvelope:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/User' }
...